Make __pagesize a size_t.
It is defined as such in libc.h, and so needs to be the same. But size_t is
defined as unsigned int on some 32-bit archs.

+++ src/init/start_libc.c
@@ -28,7 +28,7 @@ weak_alias(__init_progname, dummy2);
 
 hidden int __elevated;
 hidden int __thread_list_lock;
-hidden unsigned long __page_size;
+hidden size_t __page_size;
 hidden int __threaded;
 
 char **__environ;
